+++ src/Rocketeer/Tasks/Rollback.php
@@ -56,8 +56,9 @@ class Rollback extends AbstractTask
         }
 
         // Rollback release
-        $this->explainer->success('Rolling back to release '.$rollbackRelease);
         $this->updateSymlink($rollbackRelease);
+
+        return $this->explainer->success('Rolling back to release '.$rollbackRelease);
     }
 
     ////////////////////////////////////////////////////////////////////